basic master slave functionality working after a fashion

+++ project_code/AutoVJ.py
@@ -1,6 +1,9 @@
 #!/usr/bin/python
 from __future__ import absolute_import, division, print_function, unicode_literals
-""" A very simple gradient-colored cube and a simple solid colored box """
+""" Interactive Playful Geometries """
+
+MASTER = True # master plays the music and runs the flask server
+
 import demo
 import pi3d
 import random, time, glob, json
@@ -9,6 +12,8 @@ from subprocess import Popen, PIPE, STDOUT
 from multiprocessing import Process, Queue
 from flask import Flask, render_template, request
 
+from threading import Thread
+
 from Background import Background
 from Colors import Colors
 from Presets import geom_preset, color_preset, fields, samples
@@ -16,54 +21,66 @@ from AnimationState import AnimationState
 from SimpleCube import SimpleCube
 from ShaderTypes import ShaderTypes
 
+def slave_checker(ani_state):
+  response = urllib2.urlopen('http://192.168.1.2/update/?msg={}')
+  html = response.read()
+  msg = json.loads(html)
+  for key in msg:
+    ani_state.state[key] = msg[key]
+    
+
 counter = [None]*5
 counter[0] = 0
 
-DISPLAY = pi3d.Display.create(frames_per_second=30)
+DISPLAY = pi3d.Display.create(frames_per_second=20)
 
 ShaderTypes()
 
 mykeys = pi3d.Keyboard()
 
 perspectiveCamera = pi3d.Camera(is_3d=True)
-ortographicCamera = pi3d.Camera(is_3d=False)
 
 box = SimpleCube(perspectiveCamera)
 background = Background(perspectiveCamera)
 
 animation_state = AnimationState()
 
-#### server process and queue
-app = Flask(__name__)
-queue_down = Queue()
-queue_up = Queue()
-STATE = {}
+if MASTER:
+  #### server process and queue
+  app = Flask(__name__)
+  queue_down = Queue()
+  queue_up = Queue()
+  STATE = {}
 
-@app.route("/")
-def hello():
-  templateData = {'fields': fields} 
-  return render_template('main.html', **templateData)
+  @app.route("/")
+  def hello():
+    templateData = {'fields': fields} 
+    return render_template('main.html', **templateData)
 
-@app.route("/update/")
-def update():
-  global STATE
-  msg = request.args['msg']
-  queue_up.put(json.loads(msg))
-  if not queue_down.empty():
-    STATE = queue_down.get()
-  return json.dumps(STATE, separators=(',',':'))
+  @app.route("/update/")
+  def update():
+    global STATE
+    msg = request.args['msg']
+    queue_up.put(json.loads(msg))
+    if not queue_down.empty():
+      STATE = queue_down.get()
+    STATE['dt'] = time.time() - STATE['dt']
+    return json.dumps(STATE, separators=(',',':'))
 
-def start_server():
-  app.run(host='0.0.0.0', port=80, debug=True, use_reloader=False)
+  def start_server():
+    app.run(host='0.0.0.0', port=80, debug=False, use_reloader=False)
 
-p = Process(target=start_server)
-p.daemon = True
-p.start()
+  p = Process(target=start_server)
+  p.daemon = True
+  p.start()
 
-########### music ##########
-music = Popen(['mpg321', '-R', '-F', 'testPlayer'], stdout=PIPE, stdin=PIPE, stderr=STDOUT)
-music.stdin.write(b'LOAD music/' + animation_state.sample_progress() + b'\n')
-#############################
+  ########### music ##########
+  music = Popen(['mpg321', '-R', '-F', 'testPlayer'], stdout=PIPE, stdin=PIPE, stderr=STDOUT)
+  music.stdin.write(b'LOAD music/' + animation_state.sample_progress() + b'\n')
+  #############################
+else: ## not MASTER so SLAVE!
+  import urllib2
+  
 
 nextTime = time.time()
 chosen_geom = 0
@@ -74,72 +91,100 @@ last_frame = 0
 num_amp = 0
 av_amp = 30
 activity = 0
+next_check = 0
 
 while DISPLAY.loop_running():
-  ####### music checking #######
-  for i in range(11): # should be a better way of stopping hiccups than this
-    l = music.stdout.readline()
-    if b'@P' in l:
-      if activity > 6.0:
-        animation_state.activity = 'high'
-      elif activity > 0.1:
-        animation_state.activity = 'medium'
-      else:
-        animation_state.activity = 'low'
-      music.stdin.write(b'LOAD music/' + animation_state.sample_progress() + b'\n')
-  if b'FFT' in l: #frequency analysis
-    val_str = l.split()
-    amp = sum([int(i) for i in val_str[-3:]])
-    av_amp = 0.95 * av_amp + 0.05 * amp
-    if amp > av_amp and last_amp < av_amp:
-      num_amp += 1
-      if num_amp == 8: #8th beat, update timer
-        animation_state.state['beatf'] = int(round((animation_state.frameCount -
-                                                      last_frame) / 8.0))
-        last_frame = animation_state.frameCount
-        num_amp = 0
-        animation_state.state['light'] = 0.25 + min(0.75, av_amp / 150.0)
-        if time.time() > nextTime:
-          animation_state.beat_progress()
-    last_amp = amp
-  ##############################
   animation_state.updateTimeAndFrameCount()
-  if (animation_state.frameCount % 5) == 0:
-    #first check if anything has been sent from tablets and eat to last one
-    msg = None
-    while not queue_up.empty():
-      msg = queue_up.get()
-    if msg:
-      for mkey in msg:
-        if mkey == 'geom_jump':
-          animation_state.jumpToGeometry(msg[mkey])
-          chosen_geom = msg[mkey]
-        elif mkey == 'colr_jump':
-          animation_state.jumpToColor(msg[mkey])
-          chosen_color = msg[mkey]
-        elif mkey == 'user1':
-          animation_state.jumpToColor(0)
-          chosen_color = 0
-          animation_state.state['user1'] = [i / 255.0 for i in msg[mkey]]
-        elif mkey == 'user2':
-          animation_state.jumpToColor(0)
-          chosen_color = 0
-          animation_state.state['user2'] = [i / 255.0 for i in msg[mkey]]
+  if MASTER:
+    ####### music checking #######
+    for i in range(10):
+      """ Ideally there would be a better way of stopping hiccups than this.
+      This loop has to read all the FFTs churned out since last loop
+      which will probably depend on the sample rate but generally means
+      a readline() is needed for every 0.005s of frame time period, so for
+      15fps->13
+      20fps->10
+      30fps->7
+      This step should always limit the fps rather than Display fps which
+      should not be used
+      """
+      l = music.stdout.readline()
+      if b'@P' in l:
+        if activity > 6.0:
+          animation_state.activity = 'high'
+        elif activity > 0.1:
+          animation_state.activity = 'medium'
         else:
-          if mkey in geom_preset[chosen_geom]:
-            geom_preset[chosen_geom][mkey] = msg[mkey]
-            animation_state.jumpToGeometry(chosen_geom)
-          if mkey in color_preset[chosen_color]:
-            color_preset[chosen_color][mkey] = msg[mkey]
-            animation_state.jumpToColor(chosen_color)
-      nextTime = time.time() + 5.0
-      activity += 1.0
-      
-    activity *= 0.99
-    #clear it if nothing has consumed previous input to queue
-    while not queue_down.empty():
-      queue_down.get()
-    queue_down.put(animation_state.state)
+          animation_state.activity = 'low'
+        music.stdin.write(b'LOAD music/' + animation_state.sample_progress() + b'\n')
+    if b'FFT' in l: #frequency analysis
+      val_str = l.split()
+      amp = sum([int(i) for i in val_str[-3:]])
+      av_amp = 0.95 * av_amp + 0.05 * amp
+      if amp > av_amp and last_amp < av_amp:
+        num_amp += 1
+        if num_amp == 8: #8th beat, update timer
+          animation_state.state['beatf'] = int(round((animation_state.frameCount -
+                                                        last_frame) / 8.0))
+          last_frame = animation_state.frameCount
+          num_amp = 0
+          animation_state.state['light'] = 0.25 + min(0.75, av_amp / 150.0)
+          if time.time() > nextTime:
+            animation_state.beat_progress()
+      last_amp = amp
+    ##############################
+    if (animation_state.frameCount % 5) == 0:
+      ############ get input from tablets #########################
+      #first check if anything has been sent and eat up to last one
+      msg = None
+      while not queue_up.empty():
+        msg = queue_up.get()
+      if msg:
+        for mkey in msg:
+          if mkey == 'geom_jump':
+            animation_state.jumpToGeometry(msg[mkey])
+            chosen_geom = msg[mkey]
+          elif mkey == 'colr_jump':
+            animation_state.jumpToColor(msg[mkey])
+            chosen_color = msg[mkey]
+          elif mkey == 'user1':
+            animation_state.jumpToColor(0)
+            chosen_color = 0
+            animation_state.state['user1'] = [round(i / 255.0, 3) for i in msg[mkey]]
+          elif mkey == 'user2':
+            animation_state.jumpToColor(0)
+            chosen_color = 0
+            animation_state.state['user2'] = [round(i / 255.0, 3) for i in msg[mkey]]
+          else:
+            if mkey in geom_preset[chosen_geom]:
+              geom_preset[chosen_geom][mkey] = msg[mkey]
+              animation_state.jumpToGeometry(chosen_geom)
+            if mkey in color_preset[chosen_color]:
+              color_preset[chosen_color][mkey] = msg[mkey]
+              animation_state.jumpToColor(chosen_color)
+        nextTime = time.time() + 5.0
+        activity += 1.0
+        
+      activity *= 0.99
+      ################# send state to slaves and tablets ########
+      #clear it if nothing has consumed previous input to queue
+      while not queue_down.empty():
+        queue_down.get()
+      animation_state.state['b_rx'] = background.geometry.unif[3]
+      animation_state.state['b_ry'] = background.geometry.unif[4]
+      animation_state.state['b_rx'] = background.geometry.unif[5]
+      animation_state.state['f_rx'] = box.geometry.unif[3]
+      animation_state.state['f_ry'] = box.geometry.unif[4]
+      animation_state.state['f_rx'] = box.geometry.unif[5]
+      queue_down.put(animation_state.state)
+
+  else: ## not MASTER so SLAVE!
+    bf8 = 8 * animation_state.state['beatf']
+    if (animation_state.frameCount % bf8) == next_check:
+      t = Thread(target=slave_checker, args=(animation_state,))
+      t.start()
+      next_check = (next_check + int(animation_state.state['dt'] / 0.06) - 1) % bf8
+
   box.draw(animation_state)
   background.draw(animation_state)
 
@@ -158,6 +203,7 @@ while DISPLAY.loop_running():
 
 mykeys.close()
 DISPLAY.destroy()
-p.terminate()
-music.stdin.write(b'QUIT\n')
+if MASTER:
+  p.terminate()
+  music.stdin.write(b'QUIT\n')
